backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pyjwt==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
cryptography==41.0.7
httpx==0.25.2
passlib==1.7.4
bcrypt==4.0.1
cohere==4.46
